+++ docs/index.jsx
@@ -13,6 +13,8 @@ import ArrowsInline from '../src/components/ArrowsInline';
 import ArrowsInlineExample from '../src/components/ArrowsInline/example';
 import Badge from '../src/components/Badge';
 import BadgeExample from '../src/components/Badge/example';
+import BlockList from '../src/components/BlockList';
+import BlockListExample from '../src/components/BlockList/example';
 import Button from '../src/components/Button';
 import ButtonExample from '../src/components/Button/example';
 import ButtonRow from '../src/components/ButtonRow';
@@ -55,6 +57,10 @@ const componentsArray = [
   {
     index: Badge,
     example: BadgeExample,
+  },
+  {
+    index: BlockList,
+    example: BlockListExample,
   },
   {
     index: Button,